Create the overlay as reusable model
