+++ src/app/features/filter-bar/components/dropdown-filter-overlay/dropdown-filter-overlay.component.less
@@ -3,11 +3,6 @@
 .dropdown-overlay {
   width: 25rem;
   height: @dropdown-height;
-  padding: 1.5rem;
-  border-radius: 2rem;
-  background-color: @text-color;
-  box-shadow: 0px 0px 10px rgba(#000, 0.2);
-  overflow-y: auto;
 
   display: flex;
   flex-direction: column;
