+++ src/app/features/filter-bar/components/dropdown-filter/dropdown-filter.component.html
@@ -1,4 +1,4 @@
-<div class="dropdown" #dropdown (click)="openOverlay()">
+<div class="filter dropdown" #dropdown (click)="openOverlay()">
   <span class="title">{{ title }}</span>
   <svg class="arrow">
     <use xlink:href="#lnr-chevron-down"></use>
