+++ src/app/features/filter-bar/components/dropdown-filter/dropdown-filter.component.less
@@ -1,33 +1,7 @@
 @import "../../../../../styles/styles.less";
 
 .dropdown {
-  border-radius: 2rem;
-  border: 2px solid @secondary-color;
-  background-color: @text-color;
-  transition: all 0.2s;
-
-  cursor: pointer;
-  overflow: hidden;
-  backface-visibility: hidden;
-
-  //   padding-left: 1rem;
   width: @dropdown-width;
-  height: 3.5rem;
-  padding: 0 1rem;
-
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-
-  display: flex;
-
-  &:hover {
-    border: 2px solid darken(@secondary-color, 20%);
-  }
-
-  &:hover > .arrow {
-    fill: darken(@secondary-color, 20%);
-  }
 
   .arrow {
     width: 1.5rem;
